requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
supabase==0.7.1
requests==2.26.0
python-multipart==0.0.5
pydantic==1.8.2 